service side copy moved to _os
