+++ smbclient/__init__.py
@@ -44,6 +44,7 @@ from smbclient._os import (
     SMBStatResult,
     XATTR_CREATE,
     XATTR_REPLACE,
+    copyfile_server_side
 )
 
 try:
