+++ tests/test_smbclient_shutil.py
@@ -11,10 +11,9 @@ import pytest
 import re
 from time import sleep
 
-import smbclient
 from smbclient import open_file, mkdir, stat, symlink
 from smbclient.path import exists
-from smbclient.shutil import rmtree, copy, copytree_copy, _copyfile_server_side, copyfile
+from smbclient.shutil import rmtree, copy, copytree_copy
 
 
 def test_rmtree(smb_share):
@@ -60,7 +59,7 @@ def test_rmtree_non_existing_ignore_errors(smb_share):
     rmtree("%s\\dir2" % smb_share, ignore_errors=True)
 
 
-def test_copy2(smb_share):
+def test_copy(smb_share):
     mkdir("%s\\dir2" % smb_share)
     with open_file("%s\\file1" % smb_share, mode='w') as fd:
         fd.write(u"content")
@@ -74,7 +73,7 @@ def test_copy2(smb_share):
     assert src_stat.st_size == dst_stat.st_size
 
 
-def test_copy2_raises_when_source_and_target_identical(smb_share):
+def test_copy_raises_when_source_and_target_identical(smb_share):
     mkdir("%s\\dir2" % smb_share)
     with open_file("%s\\file1" % smb_share, mode='w') as fd:
         fd.write(u"content")
@@ -90,7 +89,7 @@ def test_copy2_raises_when_source_and_target_identical(smb_share):
         copy("%s\\file1" % smb_share, "%s\\file1" % smb_share)
 
 
-def test_copy2_with_dir_as_target(smb_share):
+def test_copy_with_dir_as_target(smb_share):
     mkdir("%s\\dir2" % smb_share)
     with open_file("%s\\file1" % smb_share, mode='w') as fd:
         fd.write(u"content")
@@ -141,69 +140,3 @@ def test_copytree_with_skip(smb_share):
 
     assert exists("%s\\dir4\\dir3\\file1" % smb_share) is True
     assert exists("%s\\dir4\\file2" % smb_share) is False
-
-
-def test_copyfile_with_same_file(smb_share):
-    with open_file("%s\\file1" % smb_share, mode='w') as fd:
-        fd.write(u"content")
-
-    if (sys.version_info > (3, 0)):
-        expected = 'are the same file'
-        context = pytest.raises(ValueError, match=re.escape(expected))
-    else:
-        context = pytest.raises(ValueError)
-
-    with context:
-        copyfile("%s\\file1" % smb_share, "%s\\file1" % smb_share)
-
-
-def test_server_side_copy(smb_share):
-    mkdir("%s\\dir2" % smb_share)
-    with open_file("%s\\file1" % smb_share, mode='w') as fd:
-        fd.write(u"content" * 1024)
-
-    sleep(0.1)
-    copy("%s\\file1" % smb_share, "%s\\dir2\\file1" % smb_share, server_side_copy=True)
-
-    src_stat = stat("%s\\file1" % smb_share)
-    dst_stat = stat("%s\\dir2\\file1" % smb_share)
-
-    assert src_stat.st_size == dst_stat.st_size
-
-
-def test_server_side_copy_across_paths_raises():
-    expected = 'Server side copy can only occur on the same drive'
-    with pytest.raises(ValueError, match=re.escape(expected)):
-        _copyfile_server_side("//host/filer1/file1", "//host/filer2/file2")
-
-
-def test_server_side_copy_target_exists(smb_share):
-    with open_file("%s\\file1" % smb_share, mode='w') as fd:
-        fd.write(u"content" * 1024)
-    with open_file("%s\\file2" % smb_share, mode='w') as fd:
-        fd.write(u"content" * 1024)
-
-    if (sys.version_info > (3, 0)):
-        expected = 'already exists'
-        context = pytest.raises(ValueError, match=re.escape(expected))
-    else:
-        context = pytest.raises(ValueError)
-
-    with context:
-        _copyfile_server_side("%s\\file1" % smb_share, "%s\\file2" % smb_share)
-
-
-def test_server_side_copy_multiple_chunks(smb_share):
-    mkdir("%s\\dir2" % smb_share)
-    with open_file("%s\\file1" % smb_share, mode='w') as fd:
-        fd.write(u"content" * 1024)
-
-    smbclient.shutil.CHUNK_SIZE = 1024
-
-    sleep(0.1)
-    copy("%s\\file1" % smb_share, "%s\\dir2\\file1" % smb_share, server_side_copy=True)
-
-    src_stat = stat("%s\\file1" % smb_share)
-    dst_stat = stat("%s\\dir2\\file1" % smb_share)
-
-    assert src_stat.st_size == dst_stat.st_size
